+++ packages/react/src/components/form/form-context.ts
@@ -1,0 +1,12 @@
+import { createContext } from 'react';
+import { FormValidationResult, Rule } from './types';
+
+interface FormContextValue {
+  validationResult?: FormValidationResult;
+  onValueChanged: (name: string, value: unknown) => void
+  onValidationRulesInitialized: (name: string, rules: Rule[]) => void
+}
+
+export const FormContext = createContext<FormContextValue | undefined>(
+  undefined,
+);
