+++ packages/react/src/components/form/types.ts
@@ -1,0 +1,8 @@
+export interface Rule {
+  validate: (value: unknown) => boolean;
+  message: string;
+}
+
+export type FormValidationResult = Record<string, string[]>;
+
+export type FormItemValidator = (value: unknown, rules: Rule[]) => string[];
